+++ src/fastrle.h
@@ -1,2 +1,7 @@
+#ifndef _MEDAKA_FASTRLE_H
+#define _MEDAKA_FASTRLE_H
+
 // Read a fasta/q and writes quality-RLE fastq file
 size_t fastrle(char *fname);
+
+#endif