+++ src/medaka_bamiter.h
@@ -1,5 +1,5 @@
-#ifndef MBAMITER_H
-#define MBAMITER_H
+#ifndef _MEDAKA_BAMITER_H
+#define _MEDAKA_BAMITER_H
 
 #include <stdbool.h>
 #include "htslib/sam.h"
@@ -13,6 +13,7 @@ typedef struct {
     char tag_name[2];
     int tag_value;
     bool keep_missing;
+    const char *read_group;
 } mplp_data;
 
 // iterator for reading bam