+++ src/medaka_counts.h
@@ -1,3 +1,6 @@
+#ifndef _MEDAKA_FASTRLE_H
+#define _MEDAKA_FASTRLE_H
+
 // medaka-style feature data
 typedef struct _plp_data {
     size_t buffer_cols;
@@ -143,5 +146,6 @@ void print_pileup_data(plp_data pileup, size_t num_dtypes, char *dtypes[], size_
 plp_data calculate_pileup(
         const char *region, const char *bam_file, size_t num_dtypes, char *dtypes[],
         size_t num_homop, const char tag_name[2], const int tag_value, const _Bool keep_missing,
-        bool weibull_summation);
+        bool weibull_summation, const char *read_group);
 
+#endif